[JIT] Rename canonicalize ops
[ghstack-poisoned]
