+++ torch/csrc/jit/passes/canonicalize_graph_fuser_ops.cpp
@@ -1,4 +1,4 @@
-#include <torch/csrc/jit/passes/canonicalize_ops.h>
+#include <torch/csrc/jit/passes/canonicalize_graph_fuser_ops.h>
 #include <torch/csrc/jit/passes/dead_code_elimination.h>
 
 namespace torch {
